requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-dotenv
passlib[bcrypt]
pyjwt
pydantic